+++ docs/release/vendors.js
@@ -1,0 +1,4 @@
+/* start:concat-scripts */
+
+/* end:concat-scripts */
+//# sourceMappingURL=vendors.js.map