+++ grunt/config/copy.js
@@ -1,0 +1,15 @@
+/**
+ * Generated header by C0ZEN for x-days-later project
+ * Generated file copy on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/01/2018
+ * Time: 14:56
+ * Version: 0.12.1
+ */
+module.exports = {
+	defaultHtml: {
+		src : 'docs/_layouts/default.tpl.html',
+		dest: 'docs/_layouts/default.useless.html'
+	}
+};