+++ grunt/config/usemin.js
@@ -1,0 +1,14 @@
+/**
+ * Generated header by Cozen for x-days-later project
+ * Generated file usemin on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/01/2018
+ * Time: 13:00
+ * Version: 0.12.1
+ */
+module.exports = {
+	html: [
+		'docs/_layouts/default.useless.html'
+	]
+};