+++ grunt/config/useminPrepare.js
@@ -1,0 +1,25 @@
+/**
+ * Generated header by Cozen for x-days-later project
+ * Generated file useminPrepare on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/01/2018
+ * Time: 14:13
+ * Version: 0.12.1
+ */
+module.exports = {
+	html   : 'docs/_layouts/default.useless.html',
+	options: {
+		dest: 'docs/release',
+		flow: {
+			steps: {
+				js : [
+					'concat'
+				],
+				css: [
+					'concat'
+				]
+			}
+		}
+	}
+};